+++ google-cloud-logging/src/main/java/com/google/cloud/logging/v2/package-info.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     https://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,57 +15,53 @@
  */
 
 /**
- * A client to Cloud Logging API.
+ * The interfaces provided are listed below, along with usage samples.
  *
- * <p>The interfaces provided are listed below, along with usage samples.
- *
- * <p>============ ConfigClient ============
- *
- * <p>Service Description: Service for configuring sinks used to route log entries.
- *
- * <p>Sample for ConfigClient:
- *
- * <pre>
- * <code>
- * try (ConfigClient configClient = ConfigClient.create()) {
- *   LogSinkName sinkName = LogSinkName.ofProjectSinkName("[PROJECT]", "[SINK]");
- *   configClient.deleteSink(sinkName);
- * }
- * </code>
- * </pre>
- *
- * ============= LoggingClient =============
+ * <p>======================= LoggingClient =======================
  *
  * <p>Service Description: Service for ingesting and querying logs.
  *
  * <p>Sample for LoggingClient:
  *
- * <pre>
- * <code>
+ * <pre>{@code
  * try (LoggingClient loggingClient = LoggingClient.create()) {
  *   LogName logName = LogName.ofProjectLogName("[PROJECT]", "[LOG]");
  *   loggingClient.deleteLog(logName);
  * }
- * </code>
- * </pre>
+ * }</pre>
  *
- * ============= MetricsClient =============
+ * <p>======================= ConfigClient =======================
+ *
+ * <p>Service Description: Service for configuring sinks used to route log entries.
+ *
+ * <p>Sample for ConfigClient:
+ *
+ * <pre>{@code
+ * try (ConfigClient configClient = ConfigClient.create()) {
+ *   GetBucketRequest request =
+ *       GetBucketRequest.newBuilder()
+ *           .setName(
+ *               LogBucketName.ofProjectLocationBucketName("[PROJECT]", "[LOCATION]", "[BUCKET]")
+ *                   .toString())
+ *           .build();
+ *   LogBucket response = configClient.getBucket(request);
+ * }
+ * }</pre>
+ *
+ * <p>======================= MetricsClient =======================
  *
  * <p>Service Description: Service for configuring logs-based metrics.
  *
  * <p>Sample for MetricsClient:
  *
- * <pre>
- * <code>
+ * <pre>{@code
  * try (MetricsClient metricsClient = MetricsClient.create()) {
  *   LogMetricName metricName = LogMetricName.of("[PROJECT]", "[METRIC]");
- *   LogMetric metric = LogMetric.newBuilder().build();
- *   LogMetric response = metricsClient.updateLogMetric(metricName, metric);
+ *   LogMetric response = metricsClient.getLogMetric(metricName);
  * }
- * </code>
- * </pre>
+ * }</pre>
  */
-@Generated("by gapic-generator")
+@Generated("by gapic-generator-java")
 package com.google.cloud.logging.v2;
 
 import javax.annotation.Generated;