+++ google-cloud-logging/src/test/java/com/google/cloud/logging/v2/MockMetricsServiceV2Impl.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     https://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.google.cloud.logging.v2;
 
 import com.google.api.core.BetaApi;
@@ -31,9 +32,10 @@ import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Queue;
+import javax.annotation.Generated;
 
-@javax.annotation.Generated("by GAPIC")
 @BetaApi
+@Generated("by gapic-generator-java")
 public class MockMetricsServiceV2Impl extends MetricsServiceV2ImplBase {
   private List<AbstractMessage> requests;
   private Queue<Object> responses;
@@ -67,75 +69,105 @@ public class MockMetricsServiceV2Impl extends MetricsServiceV2ImplBase {
   @Override
   public void listLogMetrics(
       ListLogMetricsRequest request, StreamObserver<ListLogMetricsResponse> responseObserver) {
-    Object response = responses.remove();
+    Object response = responses.poll();
     if (response instanceof ListLogMetricsResponse) {
       requests.add(request);
-      responseObserver.onNext((ListLogMetricsResponse) response);
+      responseObserver.onNext(((ListLogMetricsResponse) response));
       responseObserver.onCompleted();
     } else if (response instanceof Exception) {
-      responseObserver.onError((Exception) response);
+      responseObserver.onError(((Exception) response));
     } else {
-      responseObserver.onError(new IllegalArgumentException("Unrecognized response type"));
+      responseObserver.onError(
+          new IllegalArgumentException(
+              String.format(
+                  "Unrecognized response type %s for method ListLogMetrics, expected %s or %s",
+                  response == null ? "null" : response.getClass().getName(),
+                  ListLogMetricsResponse.class.getName(),
+                  Exception.class.getName())));
     }
   }
 
   @Override
   public void getLogMetric(
       GetLogMetricRequest request, StreamObserver<LogMetric> responseObserver) {
-    Object response = responses.remove();
+    Object response = responses.poll();
     if (response instanceof LogMetric) {
       requests.add(request);
-      responseObserver.onNext((LogMetric) response);
+      responseObserver.onNext(((LogMetric) response));
       responseObserver.onCompleted();
     } else if (response instanceof Exception) {
-      responseObserver.onError((Exception) response);
+      responseObserver.onError(((Exception) response));
     } else {
-      responseObserver.onError(new IllegalArgumentException("Unrecognized response type"));
+      responseObserver.onError(
+          new IllegalArgumentException(
+              String.format(
+                  "Unrecognized response type %s for method GetLogMetric, expected %s or %s",
+                  response == null ? "null" : response.getClass().getName(),
+                  LogMetric.class.getName(),
+                  Exception.class.getName())));
     }
   }
 
   @Override
   public void createLogMetric(
       CreateLogMetricRequest request, StreamObserver<LogMetric> responseObserver) {
-    Object response = responses.remove();
+    Object response = responses.poll();
     if (response instanceof LogMetric) {
       requests.add(request);
-      responseObserver.onNext((LogMetric) response);
+      responseObserver.onNext(((LogMetric) response));
       responseObserver.onCompleted();
     } else if (response instanceof Exception) {
-      responseObserver.onError((Exception) response);
+      responseObserver.onError(((Exception) response));
     } else {
-      responseObserver.onError(new IllegalArgumentException("Unrecognized response type"));
+      responseObserver.onError(
+          new IllegalArgumentException(
+              String.format(
+                  "Unrecognized response type %s for method CreateLogMetric, expected %s or %s",
+                  response == null ? "null" : response.getClass().getName(),
+                  LogMetric.class.getName(),
+                  Exception.class.getName())));
     }
   }
 
   @Override
   public void updateLogMetric(
       UpdateLogMetricRequest request, StreamObserver<LogMetric> responseObserver) {
-    Object response = responses.remove();
+    Object response = responses.poll();
     if (response instanceof LogMetric) {
       requests.add(request);
-      responseObserver.onNext((LogMetric) response);
+      responseObserver.onNext(((LogMetric) response));
       responseObserver.onCompleted();
     } else if (response instanceof Exception) {
-      responseObserver.onError((Exception) response);
+      responseObserver.onError(((Exception) response));
     } else {
-      responseObserver.onError(new IllegalArgumentException("Unrecognized response type"));
+      responseObserver.onError(
+          new IllegalArgumentException(
+              String.format(
+                  "Unrecognized response type %s for method UpdateLogMetric, expected %s or %s",
+                  response == null ? "null" : response.getClass().getName(),
+                  LogMetric.class.getName(),
+                  Exception.class.getName())));
     }
   }
 
   @Override
   public void deleteLogMetric(
       DeleteLogMetricRequest request, StreamObserver<Empty> responseObserver) {
-    Object response = responses.remove();
+    Object response = responses.poll();
     if (response instanceof Empty) {
       requests.add(request);
-      responseObserver.onNext((Empty) response);
+      responseObserver.onNext(((Empty) response));
       responseObserver.onCompleted();
     } else if (response instanceof Exception) {
-      responseObserver.onError((Exception) response);
+      responseObserver.onError(((Exception) response));
     } else {
-      responseObserver.onError(new IllegalArgumentException("Unrecognized response type"));
+      responseObserver.onError(
+          new IllegalArgumentException(
+              String.format(
+                  "Unrecognized response type %s for method DeleteLogMetric, expected %s or %s",
+                  response == null ? "null" : response.getClass().getName(),
+                  Empty.class.getName(),
+                  Exception.class.getName())));
     }
   }
 }